requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
simple-salesforce==1.12.9
python-dotenv==1.2.1
pandas==2.3.3
pydantic==2.12.5